requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250706
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
av==14.3.0
blake3==1.0.5
blend_modes==2.2.0
braceexpand==0.1.7
cachetools==6.1.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.9.3
contourpy==1.3.2
cupy-cuda12x==13.5.1
cycler==0.12.1
datasets==3.6.0
Deprecated==1.2.18
depyf==0.18.0
diffusers==0.34.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
easydict==1.13
einops==0.8.1
email_validator==2.2.0
fairscale==0.4.0
fastapi==0.115.12
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gguf==0.17.1
googleapis-common-protos==1.70.0
gradio==5.27.0
gradio_client==1.9.0
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.0.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lark==1.2.2
lazy_loader==0.4
liger_kernel==0.5.8
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.7.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.1.1
omegaconf==2.3.0
onnxruntime==1.21.1
openai==1.76.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.16
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post6
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytools==2025.1.2
pytorch-lightning==1.2.9
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.11
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
ruff==0.11.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
spandrel==0.4.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchmetrics==0.2.0
torchscale==0.3.0
torchsde==0.2.6
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
trampoline==0.1.2
transformers==4.53.2
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.118
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.5.post1
watchfiles==1.1.0
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
